refactor: convert action to Node.js application project
