--- packages/user-flow-gh-action/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "@user-flow-gh-action-workspace/user-flow-gh-action",
-  "version": "0.0.1",
-  "type": "commonjs",
-  "dependencies": {
-    "tslib": "^2.3.0"
-  }
-}
