--- packages/user-flow-gh-action/src/index.ts
@@ -1,3 +1,0 @@
-import { run } from './lib/main';
-
-run();
