+++ packages/user-flow-gh-action/src/main.ts
@@ -1,5 +1,5 @@
 import * as core from '@actions/core'
-import {wait} from './wait'
+import {wait} from './app/wait'
 
 export async function run(): Promise<void> {
   core.debug(`-> wait`)
