+++ tsconfig.base.json
@@ -15,9 +15,6 @@
     "skipDefaultLibCheck": true,
     "baseUrl": ".",
     "paths": {
-      "@user-flow-gh-action-workspace/user-flow-gh-action": [
-        "packages/user-flow-gh-action/src/index.ts"
-      ],
       "@user-flow-gh-action-workspace/user-flow-gh-action-e2e": [
         "packages/user-flow-gh-action-e2e/src/index.ts"
       ]
